#47 Enable GuardDuty in all regions
It is not yet possible in terraform to enable guard duty on multiple
regions with one statement.
It is also not posible to use for_each on
module providers as per the following issue: https://github.com/hashicorp/terraform/issues/24476
It is not possible to enable GuardDuty on regions disabled by default that require opt in.